RuntimeBlock: fix shared memory race condition
The RuntimeBlock member `cur_cycle` was located in shared memory and
incremented across all threads in each group. This worked as expected on
nvidia GPUs because all threads in a warp execute in lock step, ensuring
that the incremented integer was consistent. Store the instruction index
in thread-local memory and commit back to local memory at the end of
each evaluation sequence.