Evaluator: fix bug impacting branch in tail position
The prior version did not advance the index for the preallocated runtime
values array or deallocate the array for the branch not taken. Fix this
issue and add a test program covering the case. Also fix undefined
behavior where barriers were present in code that was not guaranteed to
execute for all threads in a work group.