Remove dynamic memory allocations for block groups (#44)
Support a fixed number of blocks per pass based on the maximum number of
indirect calls that may take place per pass and the maximum number of
blocks that may be reactivated per pass. The only values allocated
dynamically are the RuntimeBlock instances and the runtime values for
each instruction within the allocated blocks.

Upcoming changes will store a dynamic buffer of indirect call requests
that were not scheduled due to allocation failures or constraints on the
number of active blocks.